separate local and remote
